+++ docs/changelog.rst
@@ -8,9 +8,12 @@ v0.6.0 (in development)
 - Gave :program:`javaproperties`' :program:`format`, :program:`select`, and
   :program:`set` subcommands ``--ascii`` and ``--unicode`` options for
   controlling the escaping of non-ASCII characters in output
-- Gave :program:`properties2json` ``--ascii`` and ``--unicode`` options for
-  controlling the escaping of non-ASCII characters in output
+- Gave :program:`properties2json` and :program:`json2properties` ``--ascii``
+  and ``--unicode`` options for controlling the escaping of non-ASCII
+  characters in output
 - :program:`properties2json` now always outputs UTF-8
+- By default, comments output by :program:`json2properties` now have all
+  non-ASCII characters escaped rather than all non-Latin-1 characters
 
 v0.5.0 (2020-01-24)
 -----------------------